refactor: put all the separate functions back at the main file - unknow CORS issue
